apiext: copy compiled crds to stable file name for e2e test
The current Make build system will pump the CRD's through the
fix-crd tool and output the consolidated crds.yaml files in a folder
similiar to `yaml-3.7.0-rc.0.0.20230519193634-51fdee2e28ce-dirty.1685036041`.
This makes it tougher to write an e2e test because it doesn't have
access to the make variables that generated the unique version.

This commit adds a new Make target that will copy this file to a stable location so
that is easier to reference in e2e tests. This is primarily meant for
CI and testing as a stop-gap until we have time to improve the CRD
generation process.

Signed-off-by: Lance Austin <laustin@datawire.io>